Refactor state tracking with dataclasses
