+++ src/orca_chat/controller.py
@@ -17,8 +17,7 @@ from langchain_core.runnables.history import RunnableWithMessageHistory
 
 from orca_chat.chains import build_editor_llm, build_llm
 from orca_chat.config import LLMConfig, LLMRegistry
-from orca_chat.history import SessionState
-from orca_chat.stages import ChatStage, StageTracker
+from orca_chat.state import ChatStage, SessionState, StageTracker
 
 log = logging.getLogger(__name__)
